fixed multiple errors. refactored environment class for easier use and faster.
