+++ include/launcher.hpp
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <string>
-#include <environment.hpp>
+#include <common_classes.hpp>
 
 int launch_command(std::vector<std::string> &tokens, Environment* env);
