+++ include/parser.hpp
@@ -1,7 +1,7 @@
 #ifndef __EESH_PARSER__
 #define __EESH_PARSER__
 
-#include <environment.hpp>
+#include <common_classes.hpp>
 
 #include <vector>
 #include <string>
