+++ src/eesh.cpp
@@ -2,7 +2,7 @@
 #include <input.hpp>
 #include <vector>
 #include <launcher.hpp>
-#include <environment.hpp>
+#include <common_classes.hpp>
 #include <parser.hpp>
 #include <path.hpp>
 
@@ -12,7 +12,7 @@ int main()
 
     while (1)
     {
-        printf(env1.get_variable("PS1").value.c_str());
+        printf(env1.get_value("PS1").c_str());
         std::string input = read_input();
         std::vector<std::string> tokenized = tokenize_input(input); 
         std::vector<std::vector<std::string>>
